Update BybitService.cs ctor: check url for null and default to main url
I've got an Exception when tried to use BybitMarketDataService without specifying url. 

If you look at the ctor from BybitMarketDataService you will notice that you set URL to default null in the ctor. That means that this "null" value will be passed all the way through to the  BybitService, so your default value won't work.